LayoutTests/platform/wpe/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/radio-focus-navigation-group-first-focus-expected.txt





FAIL Both forward and backward focus navigation target the radio group first element. assert_equals: expected Element node <input type="radio" name="radio" id="a"></input> but got Element node <input type="text" id="start"></input>
